Allocation optimizations + drop secp256k1 lib + bugfixes
Using a bunch of small tricks to enhance the performance of the wallet generation + import functions, mostly for vanity purposes, and especially with v8 optimization in mind.

Also removed a dependency by swapping it fully for an existing one (huzzah to less bloat and dependency!)

Fixed a few QR code bugs, bumped year to 2022, etc!