[Wallet] Refactored CoinControl and bugfixes
The addition of a new, proper 'Coin Control' led to a nice decrease in code size / complexity to constructing TXs, now merely 3 lines for the same old process yet with better handling of errors and data. Also, some sweet bugfixes that'll patch any 'duplicate' balance issues, resolve the re-use of previously spent and confirmed TXs, and simplify the UTXO scanning code a ton with proper vout checking, less is more!