[UI] Improve unlock UX & safety
It was previously possible for a saved wallet to be 'overwritten' if a user decided to generate a new wallet rather than unlocking their old one; this was an 'under my nose' issue that I didn't realise until a community member ( thanks Flow! ) made it obvious, now unlocking is much easier and safer.