[Crypto] Updated Noble + switched signatures (#47)
* [Crypto] Upgrade to Noble 1.7.0

Does what it says on the tin, zero modifications, a 1:1 clone of noble-secp256k1 version 1.7.0, this commit was made to be easily auditable

* [Crypto] Switch MPW to Noble

Some minor changes for easier use of the Noble library, as well as rewriting Transaction Signatures to use Noble rather than a custom in-line implementation which was unaudited.

* [Crypto] Disable extraEntropy
For better Core compliance.

* [Deps] Remove deprecated secp256k1
The library was fully replaced by Noble and is no longer necessary.

* [Crypto] Fix signature concat
Fixed a bad concat with incorrect data, and refactored slightly to Duddino's suggestion.

* [Crypto] Refactor bitTrx + drop lib
This commit drops the ellipticcurve.js library in favor of Noble, and refactors the codebase accordingly, with quite a bit of reduction in code complexity, improved readability and even speed (having to take less hops through type-coercion).